refactor: headers struct and import
