+++ pysatl_tsp/c/ema_handler.c
@@ -1,4 +1,5 @@
 #include "handler.h"
+#include "ema_handler.h"
 #include <stdio.h>
 #include <stdlib.h>
